backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-genai>=1.0.0
google-cloud-texttospeech==2.17.2
google-cloud-storage==2.17.0
google-cloud-firestore==2.16.0
python-multipart==0.0.9
pillow>=11.0.0
python-dotenv==1.0.1
httpx==0.27.0
